# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_alemancenter_fiber-api_internal_repositories.ArticleView:
    properties:
      id:
        type: integer
      title:
        type: string
      visit_count:
        type: integer
    type: object
  github_com_alemancenter_fiber-api_internal_repositories.CountryRow:
    properties:
      count:
        type: integer
      country:
        type: string
    type: object
  github_com_alemancenter_fiber-api_internal_repositories.PostView:
    properties:
      id:
        type: integer
      title:
        type: string
      views:
        type: integer
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.ActiveVisitorRow:
    properties:
      browser:
        type: string
      city:
        type: string
      country:
        type: string
      current_page:
        type: string
      current_page_full:
        type: string
      ip:
        type: string
      is_member:
        type: boolean
      last_active:
        type: string
      os:
        type: string
      session_start:
        type: string
      user_agent:
        type: string
      user_email:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
  github_com_alemancenter_fiber-api_internal_services.ActivityOut:
    properties:
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      type:
        type: string
      user:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.ActivityUser'
    type: object
  github_com_alemancenter_fiber-api_internal_services.ActivityUser:
    properties:
      name:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.ChartDataRow:
    properties:
      full_date:
        type: string
      name:
        type: string
      pageViews:
        type: integer
      visitors:
        type: integer
    type: object
  github_com_alemancenter_fiber-api_internal_services.ContentAnalyticsResponse:
    properties:
      draft_articles:
        type: integer
      published_articles:
        type: integer
      top_articles:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_repositories.ArticleView'
        type: array
      top_posts:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_repositories.PostView'
        type: array
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.DashboardAnalytics:
    properties:
      articles:
        items:
          type: integer
        type: array
      authors:
        items:
          type: integer
        type: array
      comments:
        items:
          type: integer
        type: array
      dates:
        items:
          type: string
        type: array
      news:
        items:
          type: integer
        type: array
      views:
        items:
          type: integer
        type: array
    type: object
  github_com_alemancenter_fiber-api_internal_services.DashboardSummaryResponse:
    properties:
      analytics:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.DashboardAnalytics'
      onlineUsers:
        items: {}
        type: array
      recentActivities:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.ActivityOut'
        type: array
      totals:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.DashboardTotals'
      trends:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.DashboardTrends'
    type: object
  github_com_alemancenter_fiber-api_internal_services.DashboardTotals:
    properties:
      articles:
        type: integer
      news:
        type: integer
      online_users:
        type: integer
      users:
        type: integer
    type: object
  github_com_alemancenter_fiber-api_internal_services.DashboardTrends:
    properties:
      articles:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.TrendData'
      news:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.TrendData'
      users:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.TrendData'
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.DeviceStatRow:
    properties:
      color:
        type: string
      count:
        type: integer
      name:
        type: string
      value:
        type: number
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.PerformanceSummaryResponse:
    properties:
      redis_info:
        type: string
      timestamp:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.PruneAnalyticsResponse:
    properties:
      deleted:
        type: integer
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.TrafficSourceRow:
    properties:
      change:
        type: number
      source:
        type: string
      visits:
        type: integer
    type: object
  github_com_alemancenter_fiber-api_internal_services.TrendData:
    properties:
      percentage:
        type: number
      trend:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.UserStatsData:
    properties:
      active:
        type: integer
      new_today:
        type: integer
      total:
        type: integer
    type: object
  github_com_alemancenter_fiber-api_internal_services.VisitorAnalyticsResponse:
    properties:
      chart_data:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.ChartDataRow'
        type: array
      country_stats:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_repositories.CountryRow'
        type: array
      device_stats:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.DeviceStatRow'
        type: array
      traffic_sources:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.TrafficSourceRow'
        type: array
      user_stats:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.UserStatsData'
      visitor_stats:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.VisitorStatsData'
    type: object
  github_com_alemancenter_fiber-api_internal_services.VisitorStatsData:
    properties:
      active_visitors:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.ActiveVisitorRow'
        type: array
      change:
        type: number
      current:
        type: integer
      current_guests:
        type: integer
      current_members:
        type: integer
      history:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.ChartDataRow'
        type: array
      total_combined_today:
        type: integer
      total_today:
        type: integer
    type: object
  github_com_alemancenter_fiber-api_internal_utils.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      pagination:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      success:
        type: boolean
      token:
        type: string
    type: object
  github_com_alemancenter_fiber-api_internal_utils.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
  internal_handlers_ai.GenerateRequest:
    properties:
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - title
    type: object
  internal_handlers_analytics.PruneRequest:
    properties:
      days:
        type: integer
    type: object
...
host: api.alemancenter.com
info:
  contact:
    email: support@alemancenter.com
    name: API Support
  description: Backend API for Alemancenter Educational Platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Alemancenter API
  version: 2.0.0
paths:
  /ai/generate:
    post:
      consumes:
      - application/json
      description: Generate article content automatically using AI based on a provided
        title
      parameters:
      - description: Prompt payload containing title
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_ai.GenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: AI Content Generation
      tags:
      - AI
...
  /dashboard:
    get:
      description: Returns the main dashboard summary including articles, posts, and
        file counts
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.DashboardSummaryResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Dashboard Summary
      tags:
      - Analytics
...
  /dashboard/content-analytics:
    get:
      description: Get performance metrics for articles and posts (e.g., views)
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.ContentAnalyticsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Content Analytics
      tags:
      - Analytics
...
  /dashboard/performance/cache:
    get:
      description: Returns the Redis cache hit ratio and total keys
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Cache Performance
      tags:
      - Analytics
...
  /dashboard/performance/response-time:
    get:
      description: Returns the latency of an internal cache (Redis) ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: API Response Time
      tags:
      - Analytics
  /dashboard/performance/summary:
    get:
      description: Get comprehensive backend performance metrics (uptime, memory,
        GC)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.PerformanceSummaryResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Performance Summary
      tags:
      - Analytics
...
  /dashboard/visitor-analytics:
    get:
      description: Returns comprehensive visitor analytics (e.g., page views, unique
        visitors, browser stats)
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      - description: 'Number of days for analysis (default: 30)'
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.VisitorAnalyticsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Visitor Analytics
      tags:
      - Analytics
  /dashboard/visitor-analytics/prune:
    post:
      consumes:
      - application/json
      description: Delete visitor analytics data older than a specified number of
        days
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      - description: 'Days to retain (default: 90)'
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_handlers_analytics.PruneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.PruneAnalyticsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Prune Analytics
      tags:
      - Analytics
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  FrontendKeyAuth:
    description: Frontend identifier key for public endpoints.
    in: header
    name: X-Frontend-Key
    type: apiKey
swagger: "2.0"